+++ test/Unit/Managers/CampaignManagerTest.ts
@@ -1492,7 +1492,7 @@ describe('CampaignManager', () => {
         });
     });
 
-    describe('on mixed placement request', () => {
+    xdescribe('on mixed placement request', () => {
         let requestData: string = '{}';
         let assetManager: AssetManager;
         let campaignManager: CampaignManager;