+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Wed Sep 26 21:14:54 UTC 2018
+
+* Temporarily remove Mixed Placement functionality and tests[#6474](https://github.com/Applifier/unity-ads-webview/pull/6474)
+
 # Wed Sep 26 09:51:45 UTC 2018
 
 * Reporting Bad Ads AB Test Groups 11/12 [#6297](https://github.com/Applifier/unity-ads-webview/pull/6297)